Fixed bug when archive and keyword where executed in diffret os.
If the archive was created in linux os and Archive Should
Contain File keyword was executed in windows os, then the
path separator in the zip file was with backslash but Robot
Framework ${/} variable would use forward slash as folder
separator. This causes the keyword to fail in this
scenario. Fixed that paths from the archive are
normalised before doing a comparison does the archive
contain the file.